UPD Virtual connection both sides (strange behaviour) & Accepting more than 1 player
Most times when a client leaves the server instantly disconnects it and prompts a WSAECONNRESET message in the server side. And sometimes it seems not to notice and do expected behaviour that is waiting till the DISCONNECT_TIMEOUT_SECONDS is reached and then disconnects the client.
With clients seems to have no effect the programmed timeout and leaves game instantly when server is disconnected, showing the same error mentioned above.

Added random number player name generator that spins like crazy but works (No need to input a name every time)