Reliability replication base imlpemented
Now there's only the basic functionality implemented: Server sends packets of replication with a sequence number and remember them, the client gets this packets, saves the sequence numbers and in intervals sends back the ack of them.
This can be visualized in the GUI of each, and applying network losses, jitter do interfere with them and can be seen that packets gets accumulated or recieved in unnexpected order.